Update entrypoint.sh for better PaaS integration
When using a PaaS such as Jelastic, if the `/data` folder is bound to the local system, it is possible that the PaaS will put the file back here when redeploying or restarting. The `touch` command does nothing if the file already exists, and `supervisord` throw an exception: `Error: Another program is already listening on a port that one of our HTTP servers is configured to use.  Shut this program down first before starting supervisord.` By removing the file first, the problem is solved.